Fix dashboard crash — cast AVG to float in SQL
Postgres AVG() returns numeric, which Drizzle hands back as string.
.toFixed() then fails. Cast to ::float so JS gets a real number, plus
defensive Number() in the table consumer.

+++ app/(admin)/admin/page.tsx
@@ -27,7 +27,8 @@ export default async function AdminDashboardPage() {
       responseCount: count(responses.id),
       submittedCount:
         sql<number>`COUNT(CASE WHEN ${responses.status} = 'submitted' THEN 1 END)::int`,
-      avgScore: sql<number | null>`AVG(${responses.totalScore})`,
+      // Cast to float so Postgres returns a JS number (numeric → string by default).
+      avgScore: sql<number | null>`AVG(${responses.totalScore})::float`,
     })
     .from(assessments)
     .leftJoin(responses, eq(responses.assessmentId, assessments.id))
